Move ResBlock to blocks
