+++ batchflow/models/eager_torch/__init__.py
@@ -1,10 +1,10 @@
 """ Eagerly! """
 from .base import EagerTorch
 from .encoder_decoder import Encoder, Decoder, EncoderDecoder, AutoEncoder, VariationalAutoEncoder
-from .resnet import ResBlock, ResNet, ResNet18, ResNet34, ResNet50, ResNet101, ResNet152, \
+from .resnet import ResNet, ResNet18, ResNet34, ResNet50, ResNet101, ResNet152, \
                     ResNeXt18, ResNeXt34, ResNeXt50, ResNeXt101, ResNeXt152, \
                     SEResNet18, SEResNet34, SEResNet50, SEResNet101, SEResNet152, \
                     SEResNeXt18, SEResNeXt34, SEResNeXt50, SEResNeXt101, SEResNeXt152
 from .unet import UNet, ResUNet
 from .utils import get_shape, get_num_channels, get_num_dims, calc_padding, unpack_fn_from_config, safe_eval
-from .blocks import DenseBlock
+from .blocks import DenseBlock, ResBlock
